COMMUNICATION-1142 Protocol always needs to log off from the device - not always the case after Unexpected Error In case of ConnectionTimeoutException (after doing the specified number of tries), throw the error
git-svn-id: http://source.eict.vpdc/svn/labo/protocols/trunk@88828 0d108963-8a47-0410-809b-ebd9fb3b3407

+++ upl/com.energyict.mdc.upl.impl/protocolimpl-v2/src/main/java/com/energyict/protocolimplv2/nta/dsmr50/elster/am540/Dsmr50Properties.java
@@ -45,8 +45,8 @@ public class Dsmr50Properties extends DlmsProperties {
         return getProperties().getTypedProperty(AARQ_TIMEOUT_PROPERTY, BigDecimal.ZERO).longValue();
     }
 
-    public long getAARQRetries() {
-        return getProperties().getTypedProperty(AARQ_RETRIES_PROPERTY, BigDecimal.valueOf(2)).longValue();
+    public int getAARQRetries() {
+        return getProperties().getTypedProperty(AARQ_RETRIES_PROPERTY, BigDecimal.valueOf(2)).intValue();
     }
 
     @Override
